parser-vanilla: rewrite to optimize memory usage
The previous version generated and passed a lot of strings to each
function during the parsing process. To get rid of most of the temporary
variables, the entire structure had to be changed. The parser is now
using string.find instead of string.gfind and the string preparations
have moved to the place where the parsing is done for each message type.